requirements.txt
streamlit
python-dotenv
ollama
pandas
matplotlib
cryptography
hashlib-compat
supabase
